change upload video scene to be a coponent to admin scene + better file structure for admin scene
