+++ src/scenes/AdminScene/component/AddCategories.tsx
@@ -1,8 +1,8 @@
 import React, { useState, useCallback, FormEvent } from 'react';
 import { v4 as uuidV4 } from 'uuid';
-import { uploadMultipleCategory } from '../services/index';
+import { uploadMultipleCategory } from '../../../services/index';
 
-import '../styles/AdminStyle.scss';
+import '../styles/AddCategories.scss';
 
 export default function AddCategories() {
   const [categories, setCategories] = useState<
