+++ src/scenes/AdminScene/index.tsx
@@ -1,24 +1,30 @@
 import React, { useCallback, useEffect, useState, useRef } from 'react';
 import { useNavigate } from 'react-router-dom';
 
-import { generateFFmpeg, removeFFmpeg as fetchRemoveFFmpeg } from '../services/videoService';
-import { usePageTitle } from '../hooks/index';
-import AddAvatar from '../component/AddAvatar';
-import AddCategories from '../component/AddCategories';
-import { checkAuthRole } from '../services/userService';
-import { useCurrentUserContext } from '../contexts/UserAuth';
-import AddFranchise from '../component/AddFranchise';
+import { generateFFmpeg, removeFFmpeg as fetchRemoveFFmpeg } from '../../services/videoService';
+import { usePageTitle } from '../../hooks/index';
+import { useCurrentUserContext } from '../../contexts/UserAuth';
+import { checkAuthRole } from '../../services/userService';
 
-import '../styles/AdminStyle.scss';
+import MovieUpload from './component/MovieUpload';
+import CreateSeries from './component/CreateSeries';
+import EpisodeUpload from './component/EpisodeUpload';
+import AddAvatar from './component/AddAvatar';
+import AddCategories from './component/AddCategories';
+import AddFranchise from './component/AddFranchise';
+
+import './styles/index.scss';
 
 export default function AdminScene() {
   const [showAddCategories, setShowAddCategories] = useState<boolean>(false);
   const [showAddAvatar, setShowAddAvatar] = useState<boolean>(false);
   const [showFranchise, setShowFranchise] = useState<boolean>(false);
-  const [showVideo, setShowVideo] = useState<boolean>(false);
   const [showFFmpeg, setShowFFmpeg] = useState<boolean>(false);
   const [showRemoveFFmpeg, setShowRemoveFFmpeg] = useState<boolean>(false);
   const [showAddUser, setShowAddUser] = useState<boolean>(false);
+  const [showMovieUpload, setShowMovieUpload] = useState<boolean>(false);
+  const [showCreateSeries, setShowCreateSeries] = useState<boolean>(false);
+  const [showAddEpisode, setShowAddEpisode] = useState<boolean>(false);
 
   const [isUploading, setIsUploading] = useState<boolean>(false);
 
@@ -32,11 +38,8 @@ export default function AdminScene() {
   useEffect(() => setPageTitle('Admin'), [setPageTitle]);
   useEffect(() => {
     if (!currentUser || !currentUser.role) navigate('/');
-    else {
-      const run = async () =>
-        !(await checkAuthRole(currentUser.role)).data.access ? navigate('/') : null;
-      run();
-    }
+    else
+      (async () => (!(await checkAuthRole(currentUser.role)).data.access ? navigate('/') : null))();
   }, [currentUser, navigate]);
 
   const addFFmpeg = useCallback(
@@ -113,11 +116,33 @@ export default function AdminScene() {
       setShowFranchise(false);
       setShowAddUser(false);
       setShowAddAvatar(false);
-      setShowVideo(false);
       setShowAddCategories(false);
+      setShowMovieUpload(false);
+      setShowCreateSeries(false);
+      setShowAddEpisode(false);
       setValue(!value);
     },
     []
+  );
+
+  type NewType = {
+    value: boolean;
+    setValue: React.Dispatch<React.SetStateAction<boolean>>;
+    title: string;
+  };
+
+  const renderButton = useCallback(
+    ({ value, setValue, title }: NewType) => (
+      <button
+        style={{ border: value ? '1px solid #fff' : 'none' }}
+        type="button"
+        className="options-button-box"
+        onClick={() => selectField(setValue, value)}
+      >
+        <h3 className="options-box-heading">{title}</h3>
+      </button>
+    ),
+    [selectField]
   );
 
   return (
@@ -131,68 +156,43 @@ export default function AdminScene() {
         >
           <h3 className="options-box-heading">Go To Home</h3>
         </button>
-        <button
-          style={{ border: showAddCategories ? '1px solid #fff' : 'none' }}
-          type="button"
-          className="options-button-box"
-          onClick={() => selectField(setShowAddCategories, showAddCategories)}
-        >
-          <h3 className="options-box-heading">Add categories</h3>
-        </button>
-        <button
-          style={{ border: showFranchise ? '1px solid #fff' : 'none' }}
-          type="button"
-          className="options-button-box"
-          onClick={() => selectField(setShowFranchise, showFranchise)}
-        >
-          <h3 className="options-box-heading">Add Franchise</h3>
-        </button>
-        <button
-          style={{ border: showAddAvatar ? '1px solid #fff' : 'none' }}
-          type="button"
-          className="options-button-box"
-          onClick={() => selectField(setShowAddAvatar, showAddAvatar)}
-        >
-          <h3 className="options-box-heading">Add Avatars</h3>
-        </button>
-        <button
-          style={{ border: showVideo ? '1px solid #fff' : 'none' }}
-          type="button"
-          className="options-button-box"
-          onClick={() => selectField(setShowVideo, showVideo)}
-        >
-          <h3 className="options-box-heading">Add Video</h3>
-        </button>
-        <button
-          style={{ border: showFFmpeg ? '1px solid #fff' : 'none' }}
-          type="button"
-          className="options-button-box"
-          onClick={() => selectField(setShowFFmpeg, showFFmpeg)}
-        >
-          <h3 className="options-box-heading">Add ffmpeg to video</h3>
-        </button>
-        <button
-          style={{ border: showRemoveFFmpeg ? '1px solid #fff' : 'none' }}
-          type="button"
-          className="options-button-box"
-          onClick={() => selectField(setShowRemoveFFmpeg, showRemoveFFmpeg)}
-        >
-          <h3 className="options-box-heading">Remove ffmpeg to video</h3>
-        </button>
-        <button
-          style={{ border: showAddUser ? '1px solid #fff' : 'none' }}
-          type="button"
-          className="options-button-box"
-          onClick={() => selectField(setShowAddUser, showAddUser)}
-        >
-          <h3 className="options-box-heading">Add User</h3>
-        </button>
+        {renderButton({
+          value: showMovieUpload,
+          setValue: setShowMovieUpload,
+          title: 'Upload Movie',
+        })}
+        {renderButton({
+          value: showCreateSeries,
+          setValue: setShowCreateSeries,
+          title: 'Create Series',
+        })}
+        {renderButton({
+          value: showAddEpisode,
+          setValue: setShowAddEpisode,
+          title: 'Add Episodes',
+        })}
+        {renderButton({
+          value: showAddCategories,
+          setValue: setShowAddCategories,
+          title: 'Add Categories',
+        })}
+        {renderButton({ value: showFranchise, setValue: setShowFranchise, title: 'Add Franchise' })}
+        {renderButton({ value: showAddAvatar, setValue: setShowAddAvatar, title: 'Add Avatar' })}
+        {renderButton({ value: showFFmpeg, setValue: setShowFFmpeg, title: 'Add FFmpeg' })}
+        {renderButton({
+          value: showRemoveFFmpeg,
+          setValue: setShowRemoveFFmpeg,
+          title: 'Remove FFmpeg',
+        })}
+        {renderButton({ value: showAddUser, setValue: setShowAddUser, title: 'Add User' })}
       </div>
       <div className="main-section">
+        {showMovieUpload && <MovieUpload />}
+        {showCreateSeries && <CreateSeries />}
+        {showAddEpisode && <EpisodeUpload />}
         {showAddCategories && <AddCategories />}
         {showFranchise && <AddFranchise />}
         {showAddAvatar && <AddAvatar />}
-        {/* {showVideo && <AddVideo />} */}
         {/* {showAddUser && <AddUser />} */}
         {showFFmpeg && addFFmpeg()}
         {showRemoveFFmpeg && removeFFmpeg()}
