+++ src/scenes/AdminScene/styles/index.scss
@@ -1,0 +1,153 @@
+@import '../../../utils/scssVariables.scss';
+
+.admin-container {
+  display: flex;
+  flex-direction: row;
+  min-height: 100vh;
+  width: 100vw;
+  background-color: $main-color;
+  font-family: 'Roboto', sans-serif;
+  font-size: 1.5rem;
+  font-weight: 300;
+  border-radius: 5px;
+  box-shadow: 0px 0px 10px rgba(0, 0, 0, 0.5);
+
+  .options-menu {
+    display: flex;
+    align-items: center;
+    flex-direction: column;
+    width: 18%;
+    min-height: 100%;
+    background-color: #30333b;
+    font-family: 'Roboto', sans-serif;
+    font-size: 1rem;
+    font-weight: 300;
+    border-radius: 5px;
+    box-shadow: 0px 0px 10px rgba(0, 0, 0, 0.8);
+
+    .options-button-box {
+      cursor: pointer;
+      @include centerFlex;
+      flex-direction: row;
+      width: 85%;
+      height: 50px;
+      background-color: #3b3e46;
+      font-family: $Roboto-Slab, 'Roboto', sans-serif;
+      border-radius: 5px;
+      margin: 10px 0;
+      padding: 5px;
+      box-shadow: 0px 0px 10px rgba(0, 0, 0, 0.5);
+      border: none;
+
+      &:hover {
+        transition: 0.3s;
+        filter: brightness(1.2);
+      }
+
+      .options-box-heading {
+        color: #d6d6d6;
+
+        &:hover {
+          transition: 0.3s;
+          color: #ffffff;
+        }
+      }
+    }
+    #options-button-box-go-home {
+      background-color: #3a4257;
+      box-shadow: 0px 0px 10px rgba(150, 150, 150, 0.4);
+      &:hover {
+        transition: 0.3s;
+        filter: brightness(1.2);
+      }
+
+      .options-box-heading {
+        &::before {
+          content: '\2190';
+          margin: 0 10px 0 0;
+        }
+      }
+    }
+  }
+
+  .main-section {
+    display: flex;
+    flex-direction: column;
+    width: 82%;
+    min-height: 100%;
+    font-family: 'Roboto', sans-serif;
+    font-size: 1rem;
+    font-weight: 300;
+
+    .ffmpeg-container {
+      min-height: 100%;
+      @include centerFlex;
+
+      form {
+        display: flex;
+        flex-direction: column;
+
+        label {
+          margin: 10px;
+        }
+        input {
+          width: 250px;
+          height: 30px;
+          border-radius: 12px;
+          padding: 5px 15px;
+          border: 1px solid rgba(150, 150, 150, 0.7);
+
+          &:focus {
+            outline: none;
+            border: none;
+            border: 1px solid rgba(25, 25, 25, 0.9);
+          }
+        }
+        button {
+          margin: 10px 0;
+          color: $white;
+          border-radius: 8px;
+          border: 1px solid rgba(200, 200, 200, 0.5);
+          background-color: rgba(59, 65, 84, 0.9);
+          transition: all 250ms;
+          cursor: pointer;
+
+          &:hover {
+            border: 1px solid rgba(200, 200, 200, 0.5);
+            box-shadow: 0 0 10px rgba(146, 134, 134, 0.762);
+          }
+          &:disabled {
+            cursor: not-allowed;
+          }
+
+          .submit-video-button-loader {
+            display: flex;
+            justify-content: center;
+
+            &::after {
+              content: '';
+              height: 16px;
+              width: 16px;
+              border: 10px solid #dddddd;
+              border-top-color: #009579;
+              border-radius: 50%;
+              transform: rotate(0.16turn);
+              animation: loading 1s cubic-bezier(0.7, 0.9, 0.9, 0.8) infinite;
+            }
+
+            @keyframes loading {
+              /* Safari support */
+              from {
+                transform: rotate(0turn);
+              }
+
+              to {
+                transform: rotate(1turn);
+              }
+            }
+          }
+        }
+      }
+    }
+  }
+}
